requirements.txt
open3d==0.10.0
pytorch3d
plyfile
opencv-python
lpips
trimesh
tensorboard
git+https://github.com/facebookresearch/pytorch3d.git
